### 2016-05-17 #20 @whedon
Submission: Representation and Manipulation of Genomic Tuples in R
**Submitting author:** @PeteHaitch (<a href="http://orcid.org/0000-0002-8153-6258">Peter Hickey</a>)
**Repository:** <a href="https://github.com/PeteHaitch/GenomicTuples" target ="_blank">https://github.com/PeteHaitch/GenomicTuples</a>
**Version:** v1.7.3
**Archive:** <a href="http://dx.doi.org/10.5281/zenodo.53186" target ="_blank">http://dx.doi.org/10.5281/zenodo.53186</a>

- [x] **Archive:** Does the <a target="_blank" href="http://dx.doi.org/10.5281/zenodo.51498">software archive</a> resolve?

- [x] **Performance:** Have the performance claims of the software been confirmed?

- [x] **Functionality documentation:** Is the functionality of the software documented to a satisfactory level (e.g. API method documentation)?

Compiled paper PDF: [10.21105.joss.00020.pdf](https://github.com/openjournals/joss-reviews/files/268840/10.21105.joss.00020.pdf)
