### 2016-05-17 #20 @amoeba
:hand: I am reviewing this.

I'll do this tonight (after 8pm pst). if anyone is eager to do a review and wants to take this on before I get to it, please do.
