### 2016-05-18 #20 @amoeba
Interesting package. I'm completely out of my depth on the genetics part BioConductor part but I found the software easy to install and was able to run the code in the vignette. I was impressed by the detailed vignette included with the package (realizing its based on a previous vignette). My suggestions are essentially to improve the base README.md which should require copying and possibly slightly modifying existing prose from other sources such as the paper and vignette.

- A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

    Not in the README. It's unclear from either the repo subtitle or the README what the software does. If I visit the linked Bioconductor page for the stable release of the package, I can learn more. The paper also includes this information.

    Suggest making the README in this repository look something like the text at http://bioconductor.org/packages/release/bioc/html/GenomicTuples.html or include some of the text from the paper.

- Functionality: Have the functional claims of the software been confirmed?

    No claims made. I checked this as complete.

- Performance: Have the performance claims of the software been confirmed?

    No claims made. I checked this as complete.

- A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

    No (see above).

    Suggest: Expand the README with this information.

- Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

    Not in the README in the repo root. After a quick scan I don't use `@example` blocks in the source. I do, however, see a very detailed vignette, which is a standard way for R packages to provide detailed documentation for their package. I would say this counts but I would suggest the README be elaborated upon to include examples.

    Suggest: Including examples in the README, potentially taking from the vignette.

-  Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

    No on 1+2+3.

    Suggest: Edit the README to include community guidelines.

Grade: (2) Minor revisions.

Please comment in this Issue if any of my comments are confusing or if you'd like to discuss.
