### 2016-05-18 #20 @arfon
👍 thanks for the rapid review @amoeba.

> I'm completely out of my depth on the genetics part BioConductor

@amoeba - do you think we should seek a second reviewer here to evaluate the domain-specific aspects of the code?

@PeteHaitch - would you mind taking a look at @amoeba's comments and addressing them?
