### 2016-05-18 #20 @PeteHaitch
Indeed, thanks for the speedy review, @amoeba! I aim to address these comments tomorrow morning.
Briefly, however, it seems clear that JOSS would like to encourage developer's to include much of the detail in the README. I'm not against this, but for R packages, and Bioconductor packages in particular, much of this information is in the examples in the `.Rd` documentation files and the vignettes. I'm wondering whether it's worth developing some standard language to indicate this, as I see it being a common occurrence with R packages. This might avoid some redundancy and the risk of the README falling out-of-sync with the documentation proper.
