### 2016-05-18 #20 @labarba
I think the solution is simple: explain in the README this convention, including links to the appropriate locations of examples or other documentation files.
