### 2016-05-18 #20 @amoeba
@arfon I think a second reviewer could be good given that I can't fully verify the software does what it sets out to do. Asking another reviewer to simply check out the functional aspects of the code would be great.
