### 2016-05-18 #20 @PeteHaitch
@arfon: BTW, the package includes many examples beyond the vignette. However, I haven't used roxygen2 to create the `.Rd` files, hence there are no `@example` tags in the `.R` files
