### 2016-05-18 #20 @PeteHaitch
@labarba That sounds good to me, I'll do that in my amended submission.
