### 2016-05-22 7ff67cf
Response to JOSS review https://github.com/openjournals/joss-reviews/issues/20#issuecomment-219930496
