### 2016-05-22 #20 @PeteHaitch
Thanks again for the review, @amoeba. I've updated the README as follows:

- __Statement of need:__ Included intro paragraph to explain that the package is for storing and manipulating tuples of genomic loci (also added to repo subtitle). Explained that __GenomicTuples__ is intended to extend the functionality of the popular GenomicRanges package for use with tuples rather than ranges of genomic loci
- __Example usage:__ I have included a quick demo in the README, a link to the rendered vignette, and instructions on how to access the extensive documentation
- __Community guidelines:__ Added community guidelines for (1) contributing to the software, (2) reporting issues or problems, and (3) seeking support

I certainly feel your review has substantially improved the README and I believe I have addressed your concerns. I also updated the paper very slightly to improve the language
