### 2016-05-22 #20 @amoeba
This looks perfect. I ran through the install instructions and example code in the updated README.md and checked the URLs. It all looks great! Thanks for taking the time to consider my edits. You did a nice job addressing my suggestions.

I have checked off the remaining items in the review checklist.

Grade (1):  Accept

@arfon Given that edits to the source code were made after submission, does this need a new release, archive, and corresponding DOI?
