### 2016-05-22 #20 @arfon
@PeteHaitch - could you suggest a second reviewer to take a look at the library functionality that @amoeba wasn't comfortable reviewing?
