### 2016-05-22 #20 @ctb
I can take a look. I'm no R expert but I do -omic some genomes occasionally.
Will take me a day or two tho.
