### 2016-05-22 #20 @PeteHaitch
@ctb: That'd be great if you could take a look. I'm quite confident you'll be able to assess it and I don't think any super R expertise is required. If needs be, I can probably come up with some reviewers from the Bioconductor community.
