### 2016-05-26 #20 @ctb
LGTM!  I made a small pull request and would recommend that you run spellcheck on the documentation, but overall looks very complete. thx, nice work!
