### 2016-05-26 #20 @PeteHaitch
Thanks, Titus. I'll run a spell check over it tomorrow and report back here. Hopefully we'll then be good to go.
