+++ paper.bib
@@ -1,0 +1,37 @@
+@Manual{R,
+	title = {R: A Language and Environment for Statistical Computing},
+   	author = {{R Core Team}},
+    organization = {R Foundation for Statistical Computing},
+    address = {Vienna, Austria},
+    year = {2016},
+    url = {https://www.R-project.org/},
+}
+
+@Article{Bioc,
+    author = {{Huber} and {W.} and {Carey} and V. J. and {Gentleman} and {R.} and {Anders} and {S.} and {Carlson} and {M.} and {Carvalho} and B. S. and {Bravo} and H. C. and {Davis} and {S.} and {Gatto} and {L.} and {Girke} and {T.} and {Gottardo} and {R.} and {Hahne} and {F.} and {Hansen} and K. D. and {Irizarry} and R. A. and {Lawrence} and {M.} and {Love} and M. I. and {MacDonald} and {J.} and {Obenchain} and {V.} and {{Ole's}} and A. K. and {{Pag`es}} and {H.} and {Reyes} and {A.} and {Shannon} and {P.} and {Smyth} and G. K. and {Tenenbaum} and {D.} and {Waldron} and {L.} and {Morgan} and {M.}},
+    title = {{O}rchestrating high-throughput genomic analysis with {B}ioconductor},
+    journal = {Nature Methods},
+    year = {2015},
+    volume = {12},
+    number = {2},
+    pages = {115--121},
+    url = {http://www.nature.com/nmeth/journal/v12/n2/full/nmeth.3252.html},
+}
+
+@Article{GenomicRanges,
+    title = {Software for Computing and Annotating Genomic Ranges},
+    author = {Michael Lawrence and Wolfgang Huber and Herv\'e Pag\`es and Patrick Aboyoun and Marc Carlson and Robert Gentleman and Martin Morgan and Vincent Carey},
+    year = {2013},
+    journal = {{PLoS} Computational Biology},
+    volume = {9},
+    issue = {8},
+    doi = {10.1371/journal.pcbi.1003118},
+    url = {http://www.ploscompbiol.org/article/info%3Adoi%2F10.1371%2Fjournal.pcbi.1003118},
+}
+
+@Manual{SummarizedExperiment,
+    title = {SummarizedExperiment: SummarizedExperiment container},
+    author = {Martin Morgan and Valerie Obenchain and Jim Hester and Hervé Pagès},
+    year = {2016},
+    note = {R package version 1.3.1},
+}
