+++ README.md
@@ -2,9 +2,9 @@
 
 [![Bioconductor devel build status](http://bioconductor.org/shields/build/devel/bioc/GenomicTuples.svg)](http://bioconductor.org/checkResults/devel/bioc-LATEST/GenomicTuples/)
 [![Linux Build Status](https://travis-ci.org/PeteHaitch/GenomicTuples.svg?branch=master)](https://travis-ci.org/PeteHaitch/GenomicTuples)
-[![Windows Build status](https://ci.appveyor.com/api/projects/status/github/PeteHaitch/GenomicTuples?svg=true)](https://ci.appveyor.com/project/PeteHaitch/GenomicTuples)
 [![Bioconductor downloads](http://bioconductor.org/shields/downloads/GenomicTuples.svg)](http://bioconductor.org/packages/stats/bioc/GenomicTuples.html)
 [![Coverage Status](https://img.shields.io/codecov/c/github/PeteHaitch/GenomicTuples/master.svg)](https://codecov.io/github/PeteHaitch/GenomicTuples?branch=master)
+[![DOI](https://zenodo.org/badge/22085/PeteHaitch/GenomicTuples.svg)](https://zenodo.org/badge/latestdoi/22085/PeteHaitch/GenomicTuples)
 
 ---
 
