### 2016-05-26 628e2f2
Fixing bibtex entry.
From: Arfon Smith <arfon@users.noreply.github.com>

git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/GenomicTuples@117895 bc3139a8-67e5-0310-9ffc-ced21a209358
