### 2016-05-26 7e1a11b
Response to JOSS review https://github.com/openjournals/joss-reviews/issues/20#issuecomment-219930496
From: Peter Hickey <peter.hickey@gmail.com>

git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/GenomicTuples@117896 bc3139a8-67e5-0310-9ffc-ced21a209358
