+++ README.Rmd
@@ -1,0 +1,100 @@
+---
+output: github_document
+---
+  
+<!-- README.md is generated from README.Rmd. Please edit that file -->
+  
+[![Project Status: Active - The project has reached a stable, usable state and is being actively developed.](http://www.repostatus.org/badges/latest/active.svg)](http://www.repostatus.org/#active)
+[![Bioconductor devel build status](http://bioconductor.org/shields/build/devel/bioc/GenomicTuples.svg)](http://bioconductor.org/checkResults/devel/bioc-LATEST/GenomicTuples/)
+[![Linux Build Status](https://travis-ci.org/PeteHaitch/GenomicTuples.svg?branch=master)](https://travis-ci.org/PeteHaitch/GenomicTuples)
+[![Bioconductor downloads](http://bioconductor.org/shields/downloads/GenomicTuples.svg)](http://bioconductor.org/packages/stats/bioc/GenomicTuples.html)
+[![Coverage Status](https://img.shields.io/codecov/c/github/PeteHaitch/GenomicTuples/master.svg)](https://codecov.io/github/PeteHaitch/GenomicTuples?branch=master)
+[![DOI](https://zenodo.org/badge/22085/PeteHaitch/GenomicTuples.svg)](https://zenodo.org/badge/latestdoi/22085/PeteHaitch/GenomicTuples)
+
+```{r, echo = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  fig.path = "README-"
+)
+```
+
+R package: GenomicTuples
+---
+
+__GenomicTuples__ is an R/Bioconductor package that defines general purpose containers for storing and manipulating _genomic tuples_. A genomic tuple of size `m` is of the form `chromosome:strand:{pos_1, pos_2, ..., pos_m}` where `pos_1` < `pos_2` < `...` < `pos_m` are positions along the chromosome. __GenomicTuples__ aims to provide functionality for tuples of genomic co-ordinates that are analogous to those available for genomic ranges in the [__GenomicRanges__](http://bioconductor.org/packages/GenomicRanges/) R/Bioconductor package.
+
+## Installation
+
+Most users will want to install __GenomicTuples__ using the current release of Bioconductor (3.3) using:
+
+```{r, eval = FALSE}
+## try http:// if https:// URLs are not supported
+source("https://bioconductor.org/biocLite.R")
+biocLite("GenomicTuples")
+```
+
+The master branch of this repository is the development version of the package. The development version of __GenomicTuples__ can only be installed using the development version of Bioconductor (3.4). Please first read these [instructions on installing the development version of Bioconductor](http://www.bioconductor.org/developers/how-to/useDevel/); __GenomicTuples__ can then be installed by:
+
+```r
+source("http://bioconductor.org/biocLite.R")
+useDevel()
+biocLite("GenomicTuples")
+```
+
+## Quick demo
+
+Here we use the __GenomicTuples__ package to define two _GTuples_ objects, one containing 5 3-tuples and one containing 3 3-tuples, demonstrate how to identify 'equal' genomic tuples, and how this calculation would be incorrect if we were to mistakenly treat these genomic tuples as genomic ranges:
+
+```{r, eval = TRUE, echo = TRUE, message = FALSE}
+# Load the package
+library(GenomicTuples)
+
+# Create a GTuples object containing 5 3-tuples
+x <- GTuples(seqnames = c('chr1', 'chr1', 'chr1', 'chr1', 'chr2'), 
+               tuples = matrix(c(10L, 10L, 10L, 10L, 10L, 20L, 20L, 20L, 25L, 
+                                 20L, 30L, 30L, 35L, 30L, 30L), ncol = 3), 
+               strand = c('+', '-', '*', '+', '+'))
+x
[…]

[…]
