### 2016-05-26 88d5910
Fix typo, tupels -> tuples
From: C. Titus Brown <titus@idyll.org>

git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/GenomicTuples@117898 bc3139a8-67e5-0310-9ffc-ced21a209358
