### 2016-05-26 1c303ab
Spell check
From: Peter Hickey <peter.hickey@gmail.com>

git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/GenomicTuples@117899 bc3139a8-67e5-0310-9ffc-ced21a209358
