### 2016-05-26 #20 @PeteHaitch
Typos fixed and pushed
@arfon Do I need to mint a new DOI? Anything else I need do?
