### 2016-05-27 #20 @arfon
@PeteHaitch - yeah, if you've made changes then minting a new DOI would be great. Feel free to just drop the URL to the new DOI in this thread and I'll do the updates.
