### 2016-05-27 #20 @PeteHaitch
@arfon I made a [new release](https://github.com/PeteHaitch/GenomicTuples/releases) but I've gotten myself confused on the DOI and badge. Can you please help?

The [README](https://github.com/PeteHaitch/GenomicTuples/blob/master/README.md) currently uses:

```
[![DOI](https://zenodo.org/badge/22085/PeteHaitch/GenomicTuples.svg)](https://zenodo.org/badge/latestdoi/22085/PeteHaitch/GenomicTuples)
```
to display the "latest" DOI badge and link. That image currently shows __[10.5281/zenodo.51498](http://dx.doi.org/10.5281/zenodo.51498)__ but when clicked the link resolves to the latest/updated/current DOI ([__10.5281/zenodo.53186__](http://dx.doi.org/10.5281/zenodo.53186)). So I guess 2 questions:
1. Is there a lag in the badge being updated to display the latest doi?
2. Should I be using the latest doi or directly linking to to a specific DOI in my README?
Thanks
