### 2016-05-27 #20 @arfon
> 1. Is there a lag in the badge being updated to display the latest doi?

Not sure sorry. Zenodo is a third party service.

> 2. Should I be using the latest doi or directly linking to to a specific DOI in my README?

I think either is fine. The latest might make sense for people who stumble upon your repo? I'm mostly concerned that we have the correct DOI associated with this paper which is http://dx.doi.org/10.5281/zenodo.53186 right?
