### 2016-05-27 #20 @PeteHaitch
> I'm mostly concerned that we have the correct DOI associated with this paper which is http://dx.doi.org/10.5281/zenodo.53186 right?

Yep, that's the one. Thanks!
