### 2016-05-29 #20 @arfon
@PeteHaitch http://dx.doi.org/10.21105/joss.00020 :party: :rocket: :boom:

Thanks for the review @amoeba & @ctb
